Removed nanoid. Added new script to package.json. Moved the initialState state to store.js
